separate logic from component
