+++ front/src/components/CalendarBoard/index.jsx
@@ -1,24 +1,8 @@
 import React from "react";
 import { GridList } from "@material-ui/core";
 import "./style.css";
-import dayjs from "dayjs";
-import "dayjs/locale/ja";
 import CalendarElement from "../CalendaElement";
-
-dayjs.locale("ja");
-
-const createCalendar = () => {
-  // 今月の最初の日を追加
-  const firstDay = dayjs().startOf("month");
-  // 最初の日の曜日のindexを取得
-  const firstDayIndex = firstDay.day();
-
-  return Array(35).fill(0).map((_, i) => {
-    const diffFromDay = i - firstDayIndex;
-    const day = firstDay.add(diffFromDay, "day");
-    return day;
-  });
-};
+import createCalendar from "../../services/calendar";
 
 const calendar = createCalendar();
